requirements.txt
Authlib
Flask
Flask-Cors
psycopg2-binary
python-dotenv
requests
gunicorn